refactor: rework ad server parameter checking
