+++ EventHandlers/PrebidMobileMAXAdapters/Sources/MAXAdaptersError.swift
