+++ EventHandlers/PrebidMobileMAXAdapters/Sources/MAXConstants.swift
