+++ EventHandlers/PrebidMobileMAXAdapters/Sources/MAXMediationInterstitialUtils.swift
